Move the API to internal
